### 2022-01-13 #4058 @whedon
[PRE REVIEW]: SEEDPOD Ground Risk: A Python application and library for Uncrewed Aerial Systems ground risk analysis and risk-aware path finding
**Submitting author:** @aliaksei135 (<a href="http://orcid.org/0000-0003-0023-0300">Aliaksei Pilko</a>)
**Repository:** <a href="https://github.com/aliaksei135/seedpod_ground_risk/" target ="_blank">https://github.com/aliaksei135/seedpod_ground_risk/</a>
**Version:** v0.15.0
