### 2022-01-13 #4058 @danielskatz
👋 @aliaksei135 - I'll be the editor for your submission.  The next step is to find at least 2 reviewers. If you have any suggestions for potential reviewers, please mention them here in this thread (without tagging them with an @). Note that [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
