### 2022-01-13 #4058 @aliaksei135
Hi, thanks!

The software has an installer that is Windows only, however the wheels build for Linux and Windows. We have CI tests for both platforms.

I'll have a look through the list and find reviewers
