### 2022-01-17 #4058 @danielskatz
👋 @helgee & @Garethwpeters - Would one or both of you be willing to review this JOSS submission?
