### 2022-01-21 #4079 @whedon
[REVIEW]: SEEDPOD Ground Risk: A Python application and library for Uncrewed Aerial Systems ground risk analysis and risk-aware path finding
**Submitting author:** @aliaksei135 (<a href="http://orcid.org/0000-0003-0023-0300">Aliaksei Pilko</a>)
**Repository:** https://github.com/aliaksei135/seedpod_ground_risk/
**Branch with paper.md** (empty if default branch):
**Version:** v0.15.1
