### 2022-02-15 aliaksei135/seedpod_ground_risk#99 @kylebeggs
Link in README.md broken
Hello, I am a reviewer for JOSS assigned to this submission. The link to "Drone Code" in the README is broken. #openjournals/joss-reviews#4079
