### 2022-02-16 aliaksei135/seedpod_ground_risk#100 @AustinTSchaffer
Linux Installation Instructions
This repository has a `requirements-linux.txt` but doesn't seem to have explicit installation instructions for Linux. Please see below for the list of commands that I needed to run in order to install the dependencies for this project on my laptop. I already had the necessary Nvidia drivers and other libraries installed, so this list of commands is incomplete.

```bash
# Install OS package dependencies (APT instructions, I'm using Ubuntu)
# (taken from https://github.com/SciTools/cartopy/blob/main/INSTALL)
sudo apt -y install libproj-dev libgeos-dev

# Python version check
python --version
# Python 3.8.10

# setup venv
python -m venv venv
source venv/bin/activate

# Install requirements
python -m pip install -r requirements-linux.txt
```

JOSS review thread: https://github.com/openjournals/joss-reviews/issues/4079
