### 2022-02-16 aliaksei135/seedpod_ground_risk#100 @AustinTSchaffer
Would you be open to a pull request to get the project working on Linux? I'm looking to make some modifications to the `setup.py`.

Even with the commands above, I'm still not able to install or run this project, due to the `setup.py` using the contents of the Windows `requirements.txt`.
