### 2022-02-17 aliaksei135/seedpod_ground_risk#100 @aliaksei135
Thanks for reviewing!

Currently much of the explicit Windows/Linux split is due to underlying C/C++ dependencies like `GEOS` and `GDAL` that have significantly different installation requirements on different platforms, which aren't covered in the `setup.py` of the packages that depend on them (and in turn that this work depends upon).

I would very much be open to a PR that could get this working again on Linux, using `platform` to set the correct `requirements.txt` seems like the right way to go. I did at some point have this working on WSL Ubuntu, with the exception of the CUDA accelerated routines as CUDA and WSL don't get along.

For the specific case above, I suspect that `sudo apt install libbz2-dev` would do it, however there will likely be more system dependencies required.
