### 2022-02-17 7f8ed3a
Update README
Fixes #99
