### 2022-02-17 707e142
Update README
Fixes #99
