### 2022-02-17 192a4f2
Update README (#101)
Fixes #99
