### 2022-02-21 aliaksei135/seedpod_ground_risk#100 @kylebeggs
I reached where Austin stopped, installed libbz2-dev and tried to re-run and now get this
```
/home/kylebeggs/.local/lib/python3.8/site-packages/geopandas/_compat.py:106: UserWarning: The Shapely GEOS version (3.8.0-CAPI-1.13.1 ) is incompatible with the GEOS version PyGEOS was compiled with (3.8.1-CAPI-1.13.3). Conversions between both will be slow.
  warnings.warn(
Traceback (most recent call last):
  File "/home/kylebeggs/.local/bin/spgr", line 8, in <module>
    sys.exit(main())
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 829, in __call__
    return self.main(*args, **kwargs)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 782, in main
    rv = self.invoke(ctx)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 1066, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/seedpod_ground_risk/cli/spgr.py", line 88, in strike
    pop_grid = make_pop_grid(bounds, hour, resolution)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/seedpod_ground_risk/api/api.py", line 90, in make_pop_grid
    raster_shape = reproj_bounds(bounds, proj, resolution)
  File "/home/kylebeggs/.local/lib/python3.8/site-packages/seedpod_ground_risk/core/utils.py", line 38, in reproj_bounds
    raster_width = int(abs(max_x - min_x) // raster_resolution_m)
ValueError: cannot convert float NaN to integer
```
