### 2022-02-21 #4079 @aliaksei135
I really appreciate the work put into the review thus far and I am keen to get this working on Linux as well, so I would rather not discard work already done on this review.

I've partitioned out and installed Ubuntu on my laptop and will be focusing on sorting out the dependencies again in a fresh environment.
