### 2022-02-21 aliaksei135/seedpod_ground_risk#100 @AustinTSchaffer
To add to this list, end users may need to `sudo apt-get install libbz2-dev`, possibly reinstall Python afterwards if using pyenv, and run `git lfs fetch --all` and/or `git lfs checkout`. After performing those additional setup steps, I was able to get this command running locally for some constrained Lat/Lon containing Dudley, UK (chosen at random):

```
spgr map pop-density 52.480246 52.5304 -2.109753 -2.025296
```

I can't make sense of the output since I'm not a SME in this area. I've attached a screenshot of the resulting TIF.

![image](https://user-images.githubusercontent.com/16184219/155031873-98c38c21-ba52-403f-af09-f2aa7e22241c.png)
