### 2022-02-21 aliaksei135/seedpod_ground_risk#100 @aliaksei135
This is an 'issue' with the image viewer not being able to pull out specific geospatial data from the file and happens if you open the file in Windows Photos as well.

The file is a geoTIFF format which shares the same file extension as the plain TIFF image format, so most non-GIS viewers will try and open it as the plain format.

Try opening the file in a web viewer like: http://app.geotiff.io/

There should be a single alpha channel indicating the heatmap values.
