### 2022-02-22 aliaksei135/seedpod_ground_risk#100 @AustinTSchaffer
Pretty cool! I've managed to run all of the sub-commands under the `spgr map` command, results are below. Do you have any documentation for the `spgr path` command? I'm not really sure what to supply for the subcommand and args.

> Usage: spgr path [OPTIONS] COMMAND [ARGS]...
>
>   Generate paths minimising risk measures
>
> Options:
>   --help  Show this message and exit.

  <summary>(click to expand) spgr map pop-density 52.480246 52.5304 -2.109753 -2.025296</summary>

![image](https://user-images.githubusercontent.com/16184219/155041441-af254829-8e0f-4a61-b263-476a6dbaa1a1.png)

<summary>(click to expand) spgr map fatality 52.480246 52.5304 -2.109753 -2.025296</summary>

![image](https://user-images.githubusercontent.com/16184219/155041782-93d92ee3-9c84-4788-8108-358a5c448577.png)

<summary>(click to expand) spgr map strike 52.480246 52.5304 -2.109753 -2.025296</summary>

![image](https://user-images.githubusercontent.com/16184219/155041972-525125c9-2c0b-4a88-b3da-8334c1cf155c.png)
