### 2022-02-22 68785ef
Add Linux support to setup.py (#102)
