+++ setup.py
@@ -1,114 +1,31 @@
 import os
 
 from setuptools import setup, find_packages
+import platform
+import pkg_resources
 
 ####
 # setup.py is used over a static setup.cfg as dynamic determination of local wheel dependencies is required.
 ###
 
-extern_path = 'file:///' + os.path.join(os.getcwd(), 'extern', ) + os.sep
+extern_local_path = 'file:///../extern/'
+extern_full_path = 'file:///' + os.path.join(os.getcwd(), 'extern', ) + os.sep
 
-requirements = ['altgraph==0.17',
-                'attrs==20.3.0',
-                'bokeh==2.3.1',
-                f'Cartopy @ {extern_path}Cartopy-0.19.0.post1-cp38-cp38-win_amd64.whl',
-                'certifi==2020.12.5',
-                'chardet==4.0.0',
-                'click==7.1.2',
-                'click-plugins==1.1.1',
-                'cligj==0.7.1',
-                'cloudpickle==1.6.0',
-                'colorama==0.4.4',
-                'colorcet==2.0.6',
-                'commonmark==0.9.1',
-                'cycler==0.10.0',
-                'dask==2021.4.0',
-                'datashader==0.12.1',
-                'datashape==0.5.2',
-                'defusedxml==0.7.1',
-                'dill==0.3.3',
-                'distributed==2021.4.0',
-                f'fastparquet @ {extern_path}fastparquet-0.6.3-cp38-cp38-win_amd64.whl',
-                f'Fiona @ {extern_path}Fiona-1.8.19-cp38-cp38-win_amd64.whl',
-                'fsspec==2021.4.0',
-                'future==0.18.2',
-                f'GDAL @ {extern_path}GDAL-3.2.3-cp38-cp38-win_amd64.whl',
-                f'rasterio @ {extern_path}rasterio-1.2.3-cp38-cp38-win_amd64.whl',
-                'geopandas==0.9.0',
-                'geoviews==1.9.1',
-                'HeapDict==1.0.1',
-                'holoviews==1.14.3',
-                'import-profiler==0.0.3',
-                'Jinja2==2.11.3',
-                'kiwisolver==1.3.1',
-                'llvmlite==0.36.0',
-                'locket==0.2.1',
-                'Markdown==3.3.4',
-                'MarkupSafe==1.1.1',
-                'matplotlib==3.4.1',
-                'msgpack==1.0.2',
-                'multipledispatch==0.6.0',
-                'multiprocess==0.70.11.1',
-                'munch==2.5.0',
-                'numba==0.53.1',
-                'numpy==1.20.2',
-                'odfpy==1.4.1',
-                'packaging==20.9',
-                'pandas==1.2.4',
-                'panel==0.11.3',
-                'param==1.10.1',
-                'partd==1.2.0',
-                'pathos==0.2.7',
-                'pefile==2019.4.18',
-                'Pillow==8.2.0',
-                'pox==0.2.9',
-                'ppft==1.6.6.3',
-                'psutil==5.8.0',
-                'pyarrow==3.0.0',
-                'pyct==0.4.8',
-                'pygeos==0.8',
-                'Pygments==2.8.1',
-                'pyinstaller @ https://github.com/pyinstaller/pyinstaller/tarball/develop',
-                'pyinstaller-hooks-contrib==2021.1',
-                'pyparsing==2.4.7',
-                f'pyproj @ {extern_path}pyproj-3.0.1-cp38-cp38-win_amd64.whl',
-                'pyshp==2.1.3',
-                'PySide2==5.15.2',
-                'python-dateutil==2.8.1',
-                'pytz==2021.1',
-                'pyviz-comms==2.0.1',
-                'pywin32-ctypes==0.2.0',
[…]
