### 2022-02-22 aliaksei135/seedpod_ground_risk#100 @aliaksei135
Those look much better and like they should!

For `spgr path`, looks like I forgot to add the  `make` subcommand to it. I will fix.

Thanks @AustinTSchaffer for #102 , I have successfully built the wheels on a Linux box with an Xserver, instead of WSL. I will update the build instructions with more up to date commands.

I have arrived at what I think is the list of dependencies that require installing from `apt`, at least for the wheels to build and CLI functionality:
```bash
sudo apt install libgeos-dev libgdal-dev proj-bin libbz2-dev
```

Let me know if this list is missing any
