### 2022-02-22 bf15909
Update build instructions for linux platforms
