### 2022-02-23 de9ceb8
Fix shell agnostic venv activation
