### 2022-02-23 7f918c7
Update build instructions for linux platforms (#104)
* Update build instructions for linux platforms

* Fix shell agnostic venv activation
