### 2022-02-23 aliaksei135/seedpod_ground_risk#100 @kylebeggs
@aliaksei135  I am still having trouble running a command. I seem to have installed as I can run `spgr --help` with success. However, when I run the same command as Austin I get this error now:
```
> spgr map pop-density 52.480246 52.5304 -2.109753 -2.025296
/home/kylebeggs/spgr/lib/python3.8/site-packages/geopandas/_compat.py:106: UserWarning: The Shapely GEOS version (3.8.0-CAPI-1.13.1 ) is incompatible with the GEOS version PyGEOS was compiled with (3.8.1-CAPI-1.13.3). Conversions between both will be slow.
  warnings.warn(
Traceback (most recent call last):
  File "fiona/_shim.pyx", line 83, in fiona._shim.gdal_open_vector
  File "fiona/_err.pyx", line 291, in fiona._err.exc_wrap_pointer
fiona._err.CPLE_OpenFailedError: 'static_data/england_wa_2011_clipped.shp' not recognized as a supported file format.

Traceback (most recent call last):
  File "/home/kylebeggs/spgr/bin/spgr", line 8, in <module>
    sys.exit(main())
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 829, in __call__
    return self.main(*args, **kwargs)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 782, in main
    rv = self.invoke(ctx)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 1066, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/seedpod_ground_risk/cli/spgr.py", line 46, in pop_density
    raster_grid = make_pop_grid(bounds, hour, resolution)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/seedpod_ground_risk/api/api.py", line 92, in make_pop_grid
    layer.preload_data()
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/seedpod_ground_risk/layers/temporal_population_estimate_layer.py", line 33, in preload_data
    self._ingest_census_data()
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/seedpod_ground_risk/layers/temporal_population_estimate_layer.py", line 153, in _ingest_census_data
    census_wards_df = gpd.read_file(england_wa_2011_clipped_filepath()).drop(['altname', 'oldcode'], axis=1)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/geopandas/io/file.py", line 160, in _read_file
    with reader(path_or_bytes, **kwargs) as features:
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/fiona/env.py", line 408, in wrapper
    return f(*args, **kwargs)
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/fiona/__init__.py", line 256, in open
    c = Collection(path, mode, driver=driver, encoding=encoding,
  File "/home/kylebeggs/spgr/lib/python3.8/site-packages/fiona/collection.py", line 162, in __init__
    self.session.start(self, **kwargs)
  File "fiona/ogrext.pyx", line 540, in fiona.ogrext.Session.start
  File "fiona/_shim.pyx", line 90, in fiona._shim.gdal_open_vector
fiona.errors.DriverError: 'static_data/england_wa_2011_clipped.shp' not recognized as a supported file format.

```
