### 2022-02-23 aliaksei135/seedpod_ground_risk#100 @aliaksei135
@kylebeggs we had this exact issue before, in fact @Zach10a asked this on SO [here](https://stackoverflow.com/questions/68837404/fiona-not-seeing-shp-file-as-a-recognised-format)

Unfortunately a few of the data files are just above the limit, will add to build docs.
