### 2022-02-24 bfa1d74
Add instructions for LFS objects
