### 2022-03-01 #4079 @aliaksei135
@danielskatz we've made a lot of progress updating docs and packaging for Linux. To my knowledge both reviewers can now run all commands in the CLI. I assume the review is now progressing
