### 2022-03-04 aliaksei135/seedpod_ground_risk#106 @kylebeggs
Examples for using the software
Hello - can you please add some examples of using the software? A "Getting Started" section would be sufficient in the homepage README. I also don't see a reference to the API. I suppose a lot of this stuff is obvious when using the GUI, but this documentation is needed for the CLI.
Joss Review: openjournals/joss-reviews#4079
