### 2022-03-12 aliaksei135/seedpod_ground_risk#106 @Zach10a
Hi Kyle,

https://github.com/aliaksei135/seedpod_ground_risk/wiki/Use-Case-1

This can be found in the wiki. I have now updated the wiki link as it was outdated.
