+++ paper/paper.bib
@@ -3,29 +3,23 @@
 	booktitle={2017 IEEE/AIAA 36th Digital Avionics Systems Conference (DASC)},
 	title={Safety analysis paradigm for UAS: Development and use of a common architecture and fault tree model},
 	year={2017},
-	volume={},
-	number={},
 	pages={1-10},
 	doi={10.1109/DASC.2017.8102039}
 }
 @INPROCEEDINGS{Cour-Harbo:2020,
-	author={Cour-Harbo, Anders la},
+	author={la Cour-Harbo, Anders},
 	booktitle={2020 International Conference on Unmanned Aircraft Systems (ICUAS)}, 
 	title={Ground impact probability distribution for small unmanned aircraft in ballistic descent}, 
 	year={2020},
-	volume={},
-	number={},
 	pages={1442-1451},
 	doi={10.1109/ICUAS48674.2020.9213990}
 }
 
-@INBOOK{Ancel:2017,
+@INPROCEDINGS{Ancel:2017,
 	author = {Ersin Ancel and Francisco M. Capristan and John V. Foster and Ryan C. Condotta},
 	title = {Real-time Risk Assessment Framework for Unmanned Aircraft System (UAS) Traffic Management (UTM)},
 	booktitle = {17th AIAA Aviation Technology, Integration, and Operations Conference},
 	year = {2017},
-	chapter = {},
-	pages = {},
 	doi = {10.2514/6.2017-3273},
 	URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2017-3273},
 	eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2017-3273}
