### 2022-03-16 #4079 @danielskatz
@aliaksei135 - I've suggested some small changes in https://github.com/aliaksei135/seedpod_ground_risk/pull/107 - please merge this, or let me know what you disagree with, then we can proceed.
