### 2022-03-16 #4079 @aliaksei135
- Tag is [`v0.15.1`](https://github.com/aliaksei135/seedpod_ground_risk/releases/tag/v0.15.1)
- DOI is `10.5281/zenodo.6363635` [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6363635.svg)](https://doi.org/10.5281/zenodo.6363635) with updated metadata
